requirements.txt
faiss-cpu==1.13.2
fastapi==0.135.1
langchain==1.2.12
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.19
langchain-text-splitters==1.1.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pypdf==6.9.1
python-dotenv==1.2.2
uvicorn==0.42.0
psycopg2-binary
sqlalchemy
alembic
pgvector
langchain-nvidia-ai-endpoints

asyncpg

python-multipart
